Looks like completed project. Should test with some notstandart input

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: vsenniko <vsenniko@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/18 12:59:28 by vsenniko          #+#    #+#             */
-/*   Updated: 2024/09/18 15:56:32 by vsenniko         ###   ########.fr       */
+/*   Updated: 2024/09/19 19:11:45 by vsenniko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@
 int	ft_printf(const char *str, ...);
 int	write_str(char *str);
 int	ft_putnbr_decimal(int nbr, char *base);
-int	ft_putnbr_base(long nbr, char *base);
+int	ft_putnbr_base(unsigned int nbr, char *base);
 int	ft_putnbr_add(void *add, char *base);
 #endif
